backend/requirements.txt
aniso8601==7.0.0
asgiref==3.6.0
bcrypt==4.0.1
black==22.12.0
click==8.1.3
Django==3.2
django-cors-headers==3.13.0
django-filter==22.1
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
djangorestframework==3.14.0
flake8==6.0.0
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
isort==5.11.4
mccabe==0.7.0
mypy-extensions==1.0.0
pathspec==0.11.1
platformdirs==3.2.0
promise==2.3
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==1.7.1
pytz==2022.6
Rx==1.6.3
singledispatch==4.0.0
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.5.0
